# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestió de Magatzem",
        "description": "Documentació Swagger del projecte final",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/comandes": {
            "get": {
                "tags": [
                    "Comandes"
                ],
                "summary": "Llistar totes les comandes",
                "operationId": "ee27b7739d5decf364eab47f208e19ef",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comandes"
                ],
                "summary": "Crear una comanda",
                "operationId": "9d18cfe6ba11803eaf0b180f6c034dc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tipus": {
                                        "description": "purchase|sale",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    },
                                    "proveidor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comanda creada"
                    },
                    "422": {
                        "description": "Validació fallida"
                    }
                }
            }
        },
        "/api/comandes/{id}": {
            "get": {
                "tags": [
                    "Comandes"
                ],
                "summary": "Obtenir una comanda per ID",
                "operationId": "5ded15c82f5939c2d13c592e4ecb00dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comanda trobada"
                    },
                    "404": {
                        "description": "Comanda no trobada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comandes"
                ],
                "summary": "Actualitzar una comanda",
                "operationId": "4c2cdd59ded1b04cb3e0a7c1957c94d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tipus": {
                                        "description": "purchase|sale",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    },
                                    "proveidor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comanda actualitzada"
                    },
                    "404": {
                        "description": "Comanda no trobada"
                    },
                    "422": {
                        "description": "Validació fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comandes"
                ],
                "summary": "Eliminar una comanda",
                "operationId": "da2069ecaf785547fa0a45394f53ed1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comanda eliminada"
                    },
                    "404": {
                        "description": "Comanda no trobada"
                    }
                }
            }
        },
...
    },
...
}